requirements.txt
streamlit
ollama
numpy
faiss-cpu
scikit-learn
pypdf
